refactor(validators): extract orphan-node scan; retire test_no_orphan_nodes (#1207 B1)
Lift orphan_nodes/node_ids/referenced_node_ids/_excluded out of the retired legacy
validator into planner/validators/_orphan_scan.py, repoint the C008 unit importer,
and drop the coverage variant's legacy subprocess oracle. Enforcement already lives
in conventions/coverage/test_no_orphan_nodes.

Refs #1385
